refactor(uring): make up uring.hpp
