+++ include/uring/uring.hpp
@@ -41,6 +41,8 @@
 #include "uring/syscall.hpp"
 #include "uring/SQEntry.hpp"
 #include "uring/CQEntry.hpp"
+#include "uring/detail/SQ.hpp"
+#include "uring/detail/CQ.hpp"
 
 struct statx;
 
@@ -48,153 +50,25 @@ namespace liburingcxx {
 
 constexpr uint64_t LIBURING_UDATA_TIMEOUT = -1;
 
-template<unsigned URingFlags>
-class URing;
+struct URingParams final : io_uring_params {
+    /**
+     * @brief Construct a new io_uring_params without initializing
+     */
+    URingParams() noexcept = default;
 
-namespace detail {
-
-    struct URingParams final : io_uring_params {
-        /**
-         * @brief Construct a new io_uring_params without initializing
-         */
-        URingParams() noexcept = default;
-
-        /**
-         * @brief Construct a new io_uring_params with memset and flags
-         */
-        explicit URingParams(unsigned flags) noexcept {
-            memset(this, 0, sizeof(*this));
-            this->flags = flags;
-        }
-    };
-
-    class SubmissionQueue final {
-      private:
-        unsigned *khead;
-        unsigned *ktail;
-        unsigned *kring_mask;
-        unsigned *kring_entries;
-        unsigned *kflags;
-        unsigned *kdropped;
-        unsigned *array;
-        struct io_uring_sqe *sqes;
-
-        unsigned sqe_head; // memset to 0 during URing()
-        unsigned sqe_tail; // memset to 0 during URing()
-
-        size_t ring_sz;
-        void *ring_ptr;
-
-        unsigned pad[4];
-
-      private:
-        void setOffset(const io_sqring_offsets &off) noexcept {
-            khead = (unsigned *)((char *)ring_ptr + off.head);
-            ktail = (unsigned *)((char *)ring_ptr + off.tail);
-            kring_mask = (unsigned *)((char *)ring_ptr + off.ring_mask);
-            kring_entries = (unsigned *)((char *)ring_ptr + off.ring_entries);
-            kflags = (unsigned *)((char *)ring_ptr + off.flags);
-            kdropped = (unsigned *)((char *)ring_ptr + off.dropped);
-            array = (unsigned *)((char *)ring_ptr + off.array);
-        }
-
-        /**
-         * @brief Sync internal state with kernel ring state on the SQ side.
-         *
-         * @return unsigned number of pending items in the SQ ring, for the
-         * shared ring.
-         */
-        unsigned flush() noexcept {
-            const unsigned mask = *kring_mask;
-            unsigned tail = *ktail;
-            unsigned to_submit = sqe_tail - sqe_head;
-            if (to_submit == 0) return tail - *khead; // see below
-
-            /*
-             * Fill in sqes that we have queued up, adding them to the kernel
-             * ring
-             */
-            do {
-                array[tail & mask] = sqe_head & mask;
-                tail++;
-                sqe_head++;
-            } while (--to_submit);
-
-            /*
-             * Ensure that the kernel sees the SQE updates before it sees the
-             * tail update.
-             */
-            io_uring_smp_store_release(ktail, tail);
-
-            /*
-             * This _may_ look problematic, as we're not supposed to be reading
-             * SQ->head without acquire semantics. When we're in SQPOLL mode,
-             * the kernel submitter could be updating this right now. For
-             * non-SQPOLL, task itself does it, and there's no potential race.
-             * But even for SQPOLL, the load is going to be potentially
-             * out-of-date the very instant it's done, regardless or whether or
-             * not it's done atomically. Worst case, we're going to be
-             * over-estimating what we can submit. The point is, we need to be
-             * able to deal with this situation regardless of any perceived
-             * atomicity.
-             */
-            return tail - *khead;
-        }
-
-      public:
-        template<unsigned URingFlags>
-        friend class ::liburingcxx::URing;
-        SubmissionQueue() noexcept = default;
-        ~SubmissionQueue() noexcept = default;
-    };
-
-    class CompletionQueue final {
-      private:
-        unsigned *khead;
-        unsigned *ktail;
-        unsigned *kring_mask;
-        unsigned *kring_entries;
-        unsigned *kflags;
-        unsigned *koverflow;
-        struct io_uring_cqe *cqes;
-
-        size_t ring_sz;
-        void *ring_ptr;
-
-        unsigned pad[4];
-
-      private:
-        void setOffset(const io_cqring_offsets &off) noexcept {
-            khead = (unsigned *)((char *)ring_ptr + off.head);
-            ktail = (unsigned *)((char *)ring_ptr + off.tail);
-            kring_mask = (unsigned *)((char *)ring_ptr + off.ring_mask);
-            kring_entries = (unsigned *)((char *)ring_ptr + off.ring_entries);
-            if (off.flags) kflags = (unsigned *)((char *)ring_ptr + off.flags);
-            koverflow = (unsigned *)((char *)ring_ptr + off.overflow);
-            cqes = (io_uring_cqe *)((char *)ring_ptr + off.cqes);
-        }
-
-      public:
-        template<unsigned URingFlags>
-        friend class ::liburingcxx::URing;
-        CompletionQueue() noexcept = default;
-        ~CompletionQueue() noexcept = default;
-    };
-
-    struct CQEGetter {
-        unsigned submit;
-        unsigned waitNum;
-        unsigned getFlags;
-        int size;
-        void *arg;
-    };
-
-} // namespace detail
+    /**
+     * @brief Construct a new io_uring_params with memset and flags
+     */
+    explicit URingParams(unsigned flags) noexcept {
+        memset(this, 0, sizeof(*this));
+        this->flags = flags;
+    }
+};
 
 template<unsigned URingFlags>
 class [[nodiscard]] URing final {
   public:
-    using Params = detail::URingParams;
+    using Params = URingParams;
 
   private:
     using SubmissionQueue = detail::SubmissionQueue;
